requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
decorator==5.1.1
dill==0.3.9
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.55.0
ibm-cloud-sdk-core==3.22.0
ibm-platform-services==0.59.0
idna==3.10
importlib_resources==6.4.5
ipython==8.18.1
jedi==0.19.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pbr==6.1.0
pexpect==4.9.0
pillow==11.0.0
print-color==0.4.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
Pygments==2.18.0
PyJWT==2.9.0
pylatexenc==2.10
pyparsing==3.2.0
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
qiskit==1.2.4
qiskit-aer==0.15.1
qiskit-ibm-runtime==0.33.2
requests==2.32.3
requests_ntlm==1.3.0
rustworkx==0.15.1
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
stevedore==5.3.0
symengine==0.13.0
sympy==1.13.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
zipp==3.21.0
